feat(Systorage): implement check on top of logic for 'rename' method to avoid end user misuse & miss trigger

+++ pyfile/systorage.py
@@ -24,6 +24,8 @@ class Systorage(ABC):
         pass
 
     def rename(self, new_name: str) -> bool:
+        if os.path.exists(os.path.realpath(new_name)):
+            raise Exception("This method is not intended to be used as move method")
         old_path = self.__path.get_literal()
         new_path = f"{str(self.__path.get_complex().parent())}/{new_name}"
         os.rename(self.__path._literal, new_path)
